jp-mud/backend/app/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

# Create FastAPI app
app = FastAPI(
    title="JP-MUD API",
    description="API for Japanese MUD Text Adventure Game",
    version="0.1.0"
)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Adjust in production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Import and include routers
from app.api.game import router as game_router
app.include_router(game_router, prefix="/api", tags=["game"])

@app.get("/")
async def root():
    return {"message": "Welcome to JP-MUD API. See /docs for API documentation."}

if __name__ == "__main__":
    uvicorn.run(
        "app.main:app",
        host=os.getenv("HOST", "0.0.0.0"),
        port=int(os.getenv("PORT", "8020")),
        reload=True
    ) 