jp-mud/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
aiohttp==3.8.5
python-dotenv==1.0.0
loguru==0.7.0
psutil==5.9.5
requests==2.31.0 